:recycle: [Refactor]: LeftSidebar에서 세부 컴포넌트 분리
- CalendarHeader, CalendarList, UserInfo 분리
