+++ src/widgets/LeftSidebar/types.ts
@@ -1,0 +1,4 @@
+export type CalendarInfo = {
+	calendarId: number;
+	title: string;
+};
